refactor: dark mode buttons into button components
