+++ src/components/MainArea/index.tsx
@@ -1,7 +1,6 @@
 import Results from './Results'
 import Container from '@mui/material/Container'
 import ClearButton from "./ClearButton"
-import ToggleThemeButton from "./ToggleThemeButton"
 import Popup from "./Popup"
 
 
@@ -19,7 +18,6 @@ function MainArea() {
     >
       <Results />
       <ClearButton />
-      <ToggleThemeButton />
       <Popup />
     </Container>
   )
