+++ src/components/TopBar/index.tsx
@@ -3,27 +3,16 @@ import {
   IconButton,
   Toolbar,
   Box,
-  useTheme
 } from "@mui/material"
 import ClearIcon from '@mui/icons-material/Clear'
-import LightModeIcon from '@mui/icons-material/LightMode'
-import DarkModeIcon from '@mui/icons-material/DarkMode'
-import { useDispatch } from "react-redux"
-import { themeActions } from "../redux/slices/themeSlice"
+import { ToggleThemeButton } from "./Buttons"
 
 
 const Left = () => {
-  const dispatch = useDispatch()
-  const theme = useTheme()
-  const toggleMode = () => dispatch(themeActions.toggleMode())
 
   return (
     <Box sx={{ display: "flex", flex: 1, justifyContent: 'flex-start' }}>
-      <IconButton
-        onClick={toggleMode}
-      >
-        {theme.palette.mode === 'light' ? <LightModeIcon /> : <DarkModeIcon />}
-      </IconButton >
+      <ToggleThemeButton />
     </Box>
   )
 }
